Change file and dir structure
