+++ src/commands/__init__.py
@@ -1,0 +1,8 @@
+from . import address_book
+from . import notes
+from . import root
+from .completer import CommandCompleter
+from .handle import handle_command
+
+__all__ = ['address_book', 'notes', 'root',
+           'CommandCompleter', 'handle_command']
