+++ src/data/__init__.py
@@ -1,0 +1,3 @@
+from . import data_manager
+
+__all__ = ['data_manager']
