+++ src/decorators/__init__.py
@@ -1,0 +1,7 @@
+from .command_register import commands
+from .command_register import create_command_register
+from .command_register import display_help
+from .input_error import input_error
+
+__all__ = ['commands', 'create_command_register',
+           'display_help', 'input_error']
