+++ src/dtos/base.py
@@ -1,0 +1,10 @@
+class ValidationError(Exception):
+    pass
+
+
+class Field:
+    def __init__(self, value):
+        self.value = value
+
+    def __str__(self):
+        return str(self.value)
