+++ src/models/__init__.py
@@ -1,0 +1,5 @@
+from .note import Note
+from .record import DATA_TYPES
+from .record import Record
+
+__all__ = ['Note', 'Record', 'DATA_TYPES']
