+++ src/services/__init__.py
@@ -1,0 +1,5 @@
+from .address_book import AddressBook
+from .notes import NOTE_NOT_FOUND_MESSAGE
+from .notes import Notes
+
+__all__ = ['AddressBook', 'Notes', 'NOTE_NOT_FOUND_MESSAGE']
