+++ src/utils/constants.py
@@ -1,0 +1,2 @@
+DUMP_FILE = 'telium_data.pkl'
+HISTORY_FILE = '.history'
