+++ src/utils/utils.py
@@ -2,9 +2,6 @@ from abc import ABC
 from abc import abstractmethod
 from typing import List
 
-DUMP_FILE = 'telium_data.pkl'
-HISTORY_FILE = '.history'
-
 
 def parse_input(user_input):
     parts = user_input.strip().split()
